refactor: rename bundler modules for clarity and better code navigation
- Rename bundler.rs → orchestrator.rs to reflect its coordination role
- Rename hybrid_static_bundler.rs → code_generator.rs to reflect its code generation role
- Rename Bundler struct → BundleOrchestrator for public-facing API clarity
- Keep HybridStaticBundler name as internal implementation detail
- Update all imports and references throughout the codebase
- Update architecture documentation in README.md, CLAUDE.md, and AGENTS.md

This change addresses the confusing naming structure where we had both
Bundler and HybridStaticBundler, making it unclear which was the main
implementation. The new names clearly distinguish between the high-level
orchestration (BundleOrchestrator) and the low-level code generation
(HybridStaticBundler in code_generator.rs).
