+++ crates/cribo/src/code_generator.rs
