+++ crates/cribo/src/lib.rs
@@ -1,12 +1,12 @@
-pub mod bundler;
+pub mod code_generator;
 pub mod combine;
 pub mod config;
 pub mod dependency_graph;
 pub mod dirs;
-pub mod hybrid_static_bundler;
+pub mod orchestrator;
 pub mod resolver;
 pub mod unused_imports;
 pub mod util;
 
-pub use bundler::Bundler;
 pub use config::Config;
+pub use orchestrator::BundleOrchestrator;
